+++ console/src/components/Services/ApiExplorer/Rest/utils.ts
@@ -33,14 +33,19 @@ export const badgeSort = (methods: AllowedRESTMethods[]) => {
     .map(method => method.name);
 };
 
-// checkIfSubscription is a temporary method being added to prevent endpoints
+// checkIfSubscription is a method being added to prevent endpoints
 // with subscriptions being created. See Issue (#628)
 // using `any` here since 'operation' is not defined on the type DefinitionNode
 const checkIfSubscription = (queryRootNode: any) => {
   return queryRootNode?.operation === 'subscription';
 };
 
-// checkValidQuery is a helper to validate a query that's being created
+const checkIfAnonymousQuery = (queryRootNode: any) => {
+  // It probably doesn't have to be this explicit
+  return queryRootNode?.name === undefined;
+};
+
+// isQueryValid is a helper to validate a query that's being created
 // this helps avoiding to creating endpoints with comments, empty spaces
 // and queries that are subscriptions (see above comment for reference)
 export const isQueryValid = (query: string) => {
@@ -50,7 +55,20 @@ export const isQueryValid = (query: string) => {
 
   try {
     const parsedAST = parse(query);
-    return !checkIfSubscription(parsedAST.definitions[0]);
+    if (!parsedAST) {
+      return false;
+    }
+    // making sure that there's only 1 definition in the query
+    // query shouldn't be a subscription - server also throws an error for the same
+    // also a check's in place to make sure that the query is named
+    if (
+      parsedAST?.definitions?.length > 1 ||
+      checkIfSubscription(parsedAST.definitions[0]) ||
+      checkIfAnonymousQuery(parsedAST.definitions[0])
+    ) {
+      return false;
+    }
+    return true;
   } catch {
     return false;
   }