+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - server: fix inherited_roles issue when some of the underlying roles don't have permissions configured (fixes #6672)
 - server: fix action custom types failing to parse when mutually recursive
+- console: allow editing rest endpoints queries and misc ui improvements
 - cli: match ordering of keys in project metadata files with server metadata
 
 ## v2.0.0-alpha.5